\name{exprest-package}
\alias{exprest-package}
\alias{exprest}
\docType{package}
\title{
\packageTitle{exprest}
}
\description{
\packageDescription{exprest}
}
\details{
%%  ~~ An overview of how to use the package, ~~
%%  ~~ including the most important functions ~~
}
\author{
\packageAuthor{exprest}

Maintainer: \packageMaintainer{exprest}
}
\references{
%%  ~~ Literature or other references for background information ~~
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = exprest)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{exprest}}
%\section{Documentation Index}{\packageIndices{exprest}}

\seealso{
%%  ~~ Optional links to other man pages, e.g. ~~
%%  ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%%  ~~ Optional simple examples of the most important functions ~~
}
